requirements.txt
asyncpg~=0.26.0
aiohttp~=3.8.1
discord.py